add dir for api server files
