+++ api/db.js
@@ -1,0 +1,6 @@
+exports.con = require('mysql').createConnection({
+    host: 'localhost',
+    user: process.env.DB_USER,
+    password: process.env.DB_PASS,
+    database: 'kiki_saver'
+})
